+++ lib/sinatra/browse.rb
@@ -21,6 +21,7 @@ module Sinatra::Browse
     end
 
     require_relative 'browse/parameter_types/boolean'
+    require_relative 'browse/parameter_types/date_time'
     require_relative 'browse/parameter_types/float'
     require_relative 'browse/parameter_types/integer'
     require_relative 'browse/parameter_types/string'