+++ spec/shared_examples/min_max_validation.rb
@@ -8,31 +8,27 @@ shared_examples "a parameter type with min/max validation" do |options|
   max_val = options[:maximum_value].first.last
 
   describe "Minimum/Maximum value validation" do
+    extend ExampleHelpers
+
     before(:each) { get("features/#{test_route}", params) }
 
     describe "min" do
       context "with a parameter that's larger than the given minimum" do
         let(:params) { {min_key => (min_val + 5)} }
 
-        it "works fine and dandy" do
-          expect(body["min_test"]).to eq(min_val + 5)
-        end
+        it_works_fine_and_dandy
       end
 
       context "with a parameter that's exactly the given minimum" do
         let(:params) { {min_key => min_val} }
 
-        it "works fine and dandy" do
-          expect(body["min_test"]).to eq min_val
-        end
+        it_works_fine_and_dandy
       end
 
       context "with a parameter that's lower than the minimum" do
         let(:params) { {min_key => (min_val - 5)} }
 
-        it "fails with a 400 status" do
-          expect(status).to eq 400
-        end
+        it_fails_with_400_status
       end
     end
 
@@ -40,25 +36,19 @@ shared_examples "a parameter type with min/max validation" do |options|
       context "with a parameter that's lower than the maximum" do
         let(:params) { {max_key => (max_val - 5)} }
 
-        it "works fine and dandy" do
-          expect(body["max_test"]).to eq (max_val - 5)
-        end
+        it_works_fine_and_dandy
       end
 
       context "with a parameter that's exactly the given maximum" do
         let(:params) { {max_key => max_val} }
 
-        it "works fine and dandy" do
-          expect(body["max_test"]).to eq max_val
-        end
+        it_works_fine_and_dandy
       end
 
       context "with a parameter that's larger than the given maximum" do
         let(:params) { {max_key => (max_val + 5)} }
 
-        it "fails with a 400 status" do
-          expect(status).to eq 400
-        end
+        it_fails_with_400_status
       end
     end
   end